refactor(rule): 룰 관련 패키지 구조 정리
기존에 rule패키지에 너무 많은 클래스가 있었습니다.

이를 좀더 알아보기 쉽게 counter, bonus, result 세부 패키지로 나누어 패키지 구조를 정리하였습니다.
