+++ src/main/java/lotto/domain/rule/RankingCriteria.java
@@ -1,5 +1,7 @@
 package lotto.domain.rule;
 
+import lotto.domain.rule.result.MatchResult;
+
 public interface RankingCriteria {
     Rank judge(MatchResult matchResult);
 }
