+++ src/main/java/lotto/domain/rule/bonus/SingleBonusCriteria.java
@@ -1,4 +1,4 @@
-package lotto.domain.rule;
+package lotto.domain.rule.bonus;
 
 import java.util.List;
