+++ src/main/java/lotto/domain/rule/result/MatchResult.java
@@ -1,4 +1,4 @@
-package lotto.domain.rule;
+package lotto.domain.rule.result;
 
 public record MatchResult(int matchCount, boolean bonusHit) {
