+++ src/main/java/lotto/domain/rule/result/RankResult.java
@@ -1,4 +1,6 @@
-package lotto.domain.rule;
+package lotto.domain.rule.result;
+
+import lotto.domain.rule.Rank;
 
 import java.util.Comparator;
 import java.util.Map;
